% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootunif_sampling.R
\name{bootunif}
\alias{bootunif}
\title{Compute statistic with bootunif method}
\usage{
bootunif(data, statistic, R)
}
\arguments{
\item{data}{The data as a numeric vector.}

\item{statistic}{A function which when applied to data returns a vector containing the statistic(s) of interest.The first argument to statistic must be the data. No need to pass the indcices to the statistic.}

\item{R}{The number of sampling replicates. Usually this will be a single positive integer.}
}
\value{
The returned value is a list, containing the following components: \code{t0} The observed value of statistic applied to data;\code{t} A matrix with sum(R) rows each of which is a bootstrap replicate of the result of calling statistic;\code{bootunif.Statistics} A matrix like the Bootstrap Statistics.
}
\description{
Generate R replicates of a statistic applied to data.
}
\examples{
\dontrun{
x<-rnorm(19,2,3)
y<-bootunif(x,mean,1000)
meanx<-function(x,i){
x<-x[i]
mean(x)
}
z<-boot(x,meanx,1000)
y$bootunif.Statistics
z
}
}
