% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootunif_sampling.R
\name{bootunif.sample}
\alias{bootunif.sample}
\title{Bootunif Sampling}
\usage{
bootunif.sample(x)
}
\arguments{
\item{x}{the original samples with size \code{n} as a numeric vector}
}
\description{
Generate new samples based on the existing samples. Different from the Bootsrap method, we do not take the original samples, but generates new random numbers from the uniform distribution between two samples .
}
\examples{
\dontrun{
x<-rnorm(99,1,5)
y<-bootunif.sample(x)
mu=mean(y)
bias=mu-mean(x)
}
}
