% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootunif_sampling.R
\name{CVh}
\alias{CVh}
\title{Gauss leave-one-out CV}
\usage{
CVh(h, x)
}
\arguments{
\item{h}{a vector as an initial interval for bandwiths}

\item{x}{a vector of samples}
}
\value{
a number of optimal bandwith in \code{h}.
}
\description{
choose appropriate bandwidth for Gauss kernel estimator.
}
\examples{
\dontrun{
y<-sample(0:1,size=1000,replace=TRUE,prob = c(0.7,0.3))
data1<-y*rnorm(1000,0,1)
data2<-(1-y)*rnorm(1000,1,0.3)
data<-data1+data2
h<-c(0.005,0.25)
CVh(h,data)
}
}
