% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootunif_sampling.R
\name{fh.hat}
\alias{fh.hat}
\title{Gauss kernel}
\usage{
fh.hat(x, h, data)
}
\arguments{
\item{x}{the estimated value}

\item{h}{bandwith}

\item{data}{samples}
}
\value{
a vector of Gauss estimators for \code{x}
}
\description{
Kernel estimate using Gauss density function and choose appropriate bandwidth.
}
\examples{
\dontrun{
y<-sample(0:1,size=1000,replace=TRUE,prob = c(0.7,0.3))
data1<-y*rnorm(1000,0,1)
data2<-(1-y)*rnorm(1000,1,0.3)
data<-data1+data2
x0<-c(1,2.5,-0.5)
fh.hat(x0,0.1,data)
}
}
