\name{StatComp21095-package}
\alias{StatComp21095-package}
\docType{package}
\title{
  This package provides one function for sampling and one function for Gauss kernel estimate with choosing an approximate bandwith and all homeworks in class.
}
\description{
  Function bootunif.sample is used for sampling and bootunif is for estimating statistic. Function fh.hat is used to estimate sample density dunction with Gauss kernel function and CVh gives the optimal bandwith given a sequence of bandwiths.
}
\author{
Liu Dingding
Maintainer: Liu Dingding <ding1231@mail.ustc.edu.cn>
}
\keyword{ package }

